% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_adopters.R
\name{detect_adopters}
\alias{detect_adopters}
\title{Detect Adopters in the Diffusion Process}
\usage{
detect_adopters(
  num_agents,
  adj_matrix,
  final_matrix,
  original_data,
  stages,
  stages_name,
  ps_theory,
  adoption_efficacy,
  non_adoption_efficacy,
  approach,
  p_prior
)
}
\arguments{
\item{num_agents}{Numeric scalar. Total number of agents in the network.}

\item{adj_matrix}{Numeric matrix. Adjacency matrix representing the network.}

\item{final_matrix}{Numeric matrix. Additional network-related matrix (optional for this function).}

\item{original_data}{Numeric matrix. Data containing baseline and other attributes for agents.}

\item{stages}{List. A sequence of stages for the adoption process.}

\item{stages_name}{List. Names corresponding to each stage.}

\item{ps_theory}{List. Proportions of adopters for each stage.}

\item{adoption_efficacy}{List. Efficacy values added for adopters at each stage.}

\item{non_adoption_efficacy}{Numeric scalar. Efficacy value for non-adopters.}

\item{approach}{String. Centrality measure approach: "counts", "betweeness", or "closeness".}

\item{p_prior}{Numeric scalar. Proportion adjustment for subsequent stages.}
}
\value{
A list containing:
\describe{
\item{output_lst}{List of adopter indices for each stage.}
\item{update_effi}{List of updated efficacy values for each stage.}
\item{output_data}{Original data updated with new efficacy values for adopters and non-adopters.}
}
}
\description{
This function identifies adopters across multiple stages based on centrality measures, updates their efficacy,
and adjusts data for non-adopters.
}
